+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Nothing so far
+
+## [v0.3.0] - 2019-03-17
 ### Added
 - Event handler functions can now accept interfaces instead of structs
 - Add new `github.com/go-joe/joe/joetest` package for unit tests
@@ -42,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 Initial release, note that Joe is still in alpha and the API is not yet considered
 stable before the v1.0.0 release.
 
-[Unreleased]: https://github.com/go-joe/joe/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/go-joe/joe/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/go-joe/joe/compare/v0.2.0...v0.3.0
 [v0.2.0]: https://github.com/go-joe/joe/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/go-joe/joe/releases/tag/v0.1.0